+++ src/test/pythonEnvironments/discovery/locators/pipEnvHelper.functional.test.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import * as sinon from 'sinon';
 import * as platformApis from '../../../../client/common/utils/platform';
 import * as externalDependencies from '../../../../client/pythonEnvironments/common/externalDependencies';
-import { isPipenvEnvironmentRelatedToFolder } from '../../../../client/pythonEnvironments/discovery/locators/services/pipEnvHelper';
+import { isPipenvEnvironmentRelatedToFolder } from '../../../../client/pythonEnvironments/discovery/locators/services/pipenv';
 import { TEST_LAYOUT_ROOT } from '../../common/commonTestConstants';
 
 suite('Pipenv utils', () => {
